Second Commit

+++ src/main/java/kz/edu/nurym/repository/interfaces/IBankCardRepo.java
@@ -1,7 +1,6 @@
 package kz.edu.nurym.repository.interfaces;
 
 
-import kz.edu.nurym.models.Bank;
 import kz.edu.nurym.models.BankCard;
 import kz.edu.nurym.models.User;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -11,6 +10,6 @@ import java.util.Set;
 
 @Repository
 public interface IBankCardRepo extends JpaRepository<BankCard, Long> {
-    BankCard findByNumber(int number);
+    BankCard findByNumber(long number);
     Set<BankCard> findAllByUser(User user);
 }
